fix(LottoStatisticsService) : double로 형변환하여 계산한다
- totalPrize < puchaseAmount 일 때, 결과가 0이 되는 것을 방지합니다.

+++ src/main/java/lotto/service/LottoStatisticsService.java
@@ -14,7 +14,7 @@ public class LottoStatisticsService {
 
     public double calculateProfitRate(List<LottoRank> ranks, int purchaseAmount) {
         int totalPrize = sumPrize(ranks);
-        double profitRate = totalPrize / purchaseAmount * 100;
+        double profitRate = totalPrize / (double)purchaseAmount * 100;
 
         return roundToTwoDecimalPlaces(profitRate);
     }
